+++ R/vld.R
@@ -14,8 +14,12 @@
 #' vld_term(as.term(c("x[2]", "x[1]")))
 vld_term <- function(x, validate = "complete") {
   chk_string(validate)
+  if(validate == "class") {
+    deprecate_soft("0.1.0.9001", "vld_term(validate =)",
+   details = "More specifically the 'class' value of the `validate` argument has been deprecated for `vld_s3_class(class = 'term')` and the default value of validate has been replaced by 'complete'")
+  }
   chk_subset(validate, c("class", "valid", "consistent", "complete"))
-
+  
   if (!is.term(x)) {
     return(FALSE)
   }